requirements.txt
easyocr
opencv-python
Flask
pytesseract
Pillow